Fix rebase conflict and typo
