+++ src/stores/queueStore.test.ts
@@ -189,7 +189,7 @@ describe('TaskItemImpl', () => {
     })
   })
 
-  it('should parse text outptus', () => {
+  it('should parse text outputs', () => {
     const job: JobListItem = {
       ...createHistoryJob(0, 'text-job'),
       preview_output: {
